feat: create confirmation dialog on delete course

+++ src/app/courses/courses/courses.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { catchError, Observable, of } from 'rxjs';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 import { Course } from '../models/course';
 import { CoursesService } from '../services/courses.service';
 
@@ -19,6 +21,7 @@ export class CoursesComponent implements OnInit {
     private router: Router,
     private route: ActivatedRoute,
     private snackBar: MatSnackBar,
+    private readonly dialog: MatDialog
     ) {
     this.refresh()
   }
@@ -57,14 +60,22 @@ export class CoursesComponent implements OnInit {
     this.router.navigate([course.id], {relativeTo: this.route })
   }
   onDelete(course: Course) {
-    this.service.remove(course)
-      .subscribe((data) => {
-        this.refresh();
-        if (data) {
-          this.onSucess('Curso removido com sucesso')
-          return
-        }
-        this.onError('Não foi possível remover curso')
-      });
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: "Deseja remover este curso?",
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.service.remove(course)
+          .subscribe((data) => {
+            this.refresh();
+            if (data) {
+              this.onSucess('Curso removido com sucesso')
+              return
+            }
+            this.onError('Não foi possível remover curso')
+          });
+      }
+    });
   }
 }
